+++ repos/cache/src/commonMain/kotlin/dev/inmo/micro_utils/repos/cache/cache/KVCache.kt
@@ -2,5 +2,6 @@ package dev.inmo.micro_utils.repos.cache.cache
 
 import dev.inmo.micro_utils.repos.*
 
-interface KVCache<K, V> : KeyValueRepo<K, V>
-
+interface KVCache<K, V> : KeyValueRepo<K, V> {
+    companion object
+}